회원탈퇴 구현

+++ src/main/java/com/koreate/betty/domain/member/controller/MemberApiController.java
@@ -1,15 +1,33 @@
 package com.koreate.betty.domain.member.controller;
 
+import java.util.HashMap;
+import java.util.Map;
+
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.util.WebUtils;
 
-import com.koreate.betty.domain.member.dto.form.UpdateForm;
+import com.koreate.betty.domain.member.exception.MessageException;
+import com.koreate.betty.domain.member.exception.NotFoundIdException;
 import com.koreate.betty.domain.member.service.MemberService;
+import com.koreate.betty.domain.member.service.SignService;
+import com.koreate.betty.domain.member.util.Base64Util;
 import com.koreate.betty.domain.member.vo.Inquiry;
+import com.koreate.betty.domain.member.vo.Member;
+import com.koreate.betty.domain.model.CookieConst;
+import com.koreate.betty.domain.model.SessionConst;
+import com.koreate.betty.global.error.ErrorResult;
 import com.koreate.betty.infra.email.EmailSender;
 
 import lombok.RequiredArgsConstructor;
@@ -23,24 +41,53 @@ public class MemberApiController {
 	
 	private final EmailSender emailSender;
 	private final MemberService memberService;
+	private final SignService SignService;
 	
 	@PostMapping("inquiry")
-	public String memberSendMail(Inquiry inquiry) {
+	public boolean memberSendMail(@PathVariable String id, Inquiry inquiry) {
 		if (inquiry == null) {
 			throw new RuntimeException("요청 오류");
 		}
 		emailSender.inquiry(inquiry);
-		return "redirect:/members/{memberId}";
+		return true;
+	}
+	
+	@ExceptionHandler
+	@ResponseStatus(HttpStatus.BAD_GATEWAY)
+	public ErrorResult mailHandle(MessageException ex) {
+		log.error("error = {}",ex.getMessage());
+		return new ErrorResult(HttpStatus.BAD_GATEWAY.value(), "메시지 전송 실패");
 	}
 	
 	@DeleteMapping
-	public String deleteMember(@PathVariable String id, RedirectAttributes rttr) {
-		int result = memberService.deleteMember(id);
-		if(result !=0) {
-			return "redirect:/";
+	public Map<String,String> deleteMember(String id, @PathVariable("id") String userId, HttpServletRequest request, HttpServletResponse response) {
+		int result = 0;
+		log.info("id = {}, userId = {}",id,userId);
+		if(userId.equals(id)) {
+			result = memberService.deleteMember(id); 
+			HttpSession session = request.getSession(false);
+			Member user = (Member) session.getAttribute(SessionConst.USER);
+			session.invalidate();
+			Cookie findCookie = WebUtils.getCookie(request, CookieConst.COOKIE_USER);
+			if (findCookie != null) {
+				String encodedId = Base64Util.encode(user.getId());
+				Cookie cookie = new Cookie(CookieConst.COOKIE_USER, encodedId);
+				cookie.setMaxAge(0);
+				cookie.setPath("/");
+				response.addCookie(cookie);
+			}
+			Map<String, String> map = new HashMap<>();
+			map.put("code","success");
+			return map;
 		}
-		rttr.addFlashAttribute("message", "삭제 실패");
-		return "redirect:/members/"+id;
+		throw new NotFoundIdException();
+	}
+	
+	@ExceptionHandler
+	@ResponseStatus(HttpStatus.BAD_REQUEST)
+	public ErrorResult mailHandle(NotFoundIdException ex) {
+		log.error("error = {}",ex.getMessage());
+		return new ErrorResult(HttpStatus.BAD_REQUEST.value(), "아이디 검색 실패");
 	}
